newest code

+++ Backend/routes/departmentRouter.js
@@ -1,5 +1,6 @@
 import express from "express";
 import Department from "../Model/departmentModel.js";
+import User from "../Model/userModel.js";
 import expressAsyncHandler from "express-async-handler";
 import { isAdmin, isAuth } from "../utils.js";
 
@@ -7,8 +8,6 @@ const departmentRouter = express.Router();
 
 departmentRouter.get(
   "/",
-  isAuth, 
-  isAdmin,
   expressAsyncHandler(async (req, res) => {
     const departments = await Department.find({});
     res.send(departments);
@@ -17,8 +16,6 @@ departmentRouter.get(
 
 departmentRouter.post(
   "/createDepartment",
-  isAuth, 
-  isAdmin,
   expressAsyncHandler(async (req, res) => {
     const newDepartment = new Department({
       name: req.body.name,
@@ -31,15 +28,34 @@ departmentRouter.post(
   })
 );
 
+// departmentRouter.delete(
+//   "/:id",
+//   expressAsyncHandler(async (req, res) => {
+//     const department = await Department.findById(req.params.id);
+//     if (department) {
+//       await department.deleteOne();
+//       res.send({ message: "Department Deleted" });
+//     } else {
+//       res.status(404).send({ message: "Department Not Found" });
+//     }
+//   })
+// );
 departmentRouter.delete(
   "/:id",
+  isAuth,
   expressAsyncHandler(async (req, res) => {
-    const department = await Department.findById(req.params.id);
-    if (department) {
-      await department.deleteOne();
-      res.send({ message: "Department Deleted" });
-    } else {
-      res.status(404).send({ message: "Department Not Found" });
+    const users = await User.find({ department: req.params.id });
+    if(users.length === 0){
+      const department = await Department.findById(req.params.id);
+      if (department) {
+        await department.deleteOne();
+        res.send({ message: "Department Deleted" });
+      } else {
+        res.status(404).send({ message: "Department Not Found" });
+      }
+    }
+    else{
+      res.status(404).send({ message: "This department can be deleted" });
     }
   })
 );
